Replace old pdf generation
