+++ src/components/rangeUsePlanPage/pdf/pdf/IndicatorPlants.js
