+++ src/components/rangeUsePlanPage/pdf/pdf/PlantCommunity.js
@@ -2,7 +2,7 @@ import React from 'react'
 import { StyleSheet, View } from '@react-pdf/renderer'
 import Field from './common/Field'
 import { config } from './common/config'
-import { capitalize } from '../../pdf/helper'
+import { capitalize } from '../helper'
 import Row from './common/Row'
 import SectionHeader from './common/SectionHeader'
 import MonitoringArea from './MonitoringArea'
