+++ src/components/rangeUsePlanPage/pdf/pdf/Schedules.js
@@ -7,7 +7,7 @@ import { config } from './common/config'
 import Table from './common/Table'
 import Subtext from './common/Subtext'
 import moment from 'moment'
-import { formatGrazingSchedules } from '../../pdf/helper'
+import { formatGrazingSchedules } from '../helper'
 import SectionHeader from './common/SectionHeader'
 import Row from './common/Row'
 import Field from './common/Field'
